etl/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
elasticsearch==7.9.1
fastapi==0.61.1
frozenlist==1.3.0
h11==0.13.0
hiredis==2.0.0
idna==3.3
multidict==6.0.2
orjson==3.4.1
psycopg2-binary==2.9
pydantic==1.9.0
python-dotenv==0.19.2
starlette==0.13.6
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.12.2
uvloop==0.14.0
yarl==1.7.2
